requirements.txt
pyinstaller==5.10.1
pillow==10.4.0
qrcode[pil]==7.4.2
python-barcode==0.14.0
pyperclip==1.8.2
requests==2.31.0
cryptography==41.0.4